make model selector a drawer on mobile
